requirements/test.txt
-r _base.txt
pytest-django
coverage
selenium
faker
factory_boy